updated FileController

+++ src/controller/FileController.ts
@@ -3,7 +3,7 @@ import {NextFunction, Request, Response} from "express";
 import {StatusCodes} from "http-status-codes";
 import {FileDto} from "@models/Dto/FileDto";
 import {File} from "@models/File";
-import {writeFile, unlink} from "node:fs";
+import {unlink, writeFile} from "node:fs";
 import * as path from "path";
 
 export default class FileController {
@@ -23,7 +23,7 @@ export default class FileController {
 		try {
 			const recipe = await FileController.FILE_REPOSITORY.getFilesById(Number(req.params.id));
 			if (recipe) {
-				res.json(recipe.toDto())
+				res.json(recipe.toDto());
 			}
 			else {
 				res.sendStatus(StatusCodes.NOT_FOUND);
@@ -53,19 +53,22 @@ export default class FileController {
 				res.status(StatusCodes.BAD_REQUEST).json({
 					"message": "Can not create file without content"
 				})
-				return next();
+				return;
 			}
 
 			const filePath = path.join("files", "covers", `${new Date().getTime()}-${fileDto.name}`);
 			const buffer = Buffer.from(blob, 'base64');
 			writeFile(path.join(FileController.FILES_PATH, filePath), buffer, async function (error) {
 				if (error) {
-					return next(error)
-				} else {
+					next(error);
+					return;
+				}
+				else {
 					const file = new File(fileDto.name, fileDto.type, filePath, fileDto.size)
 					const createdFile = await FileController.FILE_REPOSITORY.createFile(file);
 
-					return res.status(StatusCodes.CREATED).json(createdFile.toDto());
+					res.status(StatusCodes.CREATED).json(createdFile.toDto());
+					return;
 				}
 			});
 
@@ -77,10 +80,21 @@ export default class FileController {
 	public static async deleteFile(req: Request, res: Response, next: NextFunction) {
 		try {
 			const fileId = Number(req.params.id);
-			const file = await FileController.FILE_REPOSITORY.deleteFile(fileId);
-			console.log(file); // TODO delete file from fs
+			const file = await FileController.FILE_REPOSITORY.getFilesById(fileId);
+			if (!file) {
+				res.sendStatus(StatusCodes.NOT_FOUND);
+				return;
+			}
 
-			res.sendStatus(StatusCodes.OK)
+			await FileController.FILE_REPOSITORY.deleteFile(fileId);
+			unlink(`./${file.path}`, (err) => {
+				if (err) {
+					next(err.message);
+					return;
+				}
+
+				res.sendStatus(StatusCodes.OK)
+			})
 		} catch (error: any) {
 			next(error);
 		}
